+++ door/asyncio2.py
@@ -1,14 +1,14 @@
 """:mod:`door.asyncio2` defines utilities for asynchronous programming."""
 
-from asyncio import Lock
+from asyncio import Condition, Lock
 from dataclasses import dataclass, field
 
-from door.primitives import Acquirable, AsyncSLock
+from door.primitives import Acquirable, AsyncRSLock, AsyncWSLock, Waitable
 
 
 @dataclass
-class SLock(AsyncSLock):
-    """The class for shared locks.
+class RSLock(AsyncRSLock):
+    """The class for asynchronous read-preferring shared locks.
 
     This class is designed to be used for asynchronous programming.
 
@@ -19,3 +19,13 @@ class SLock(AsyncSLock):
 
     _r: Acquirable = field(default_factory=Lock, init=False)
     _g: Acquirable = field(default_factory=Lock, init=False)
+
+
+@dataclass
+class WSLock(AsyncWSLock):
+    """The class for asynchronous write-preferring shared locks.
+
+    This class is designed to be used for asynchronous programming.
+    """
+
+    _g: Waitable = field(default_factory=Condition, init=False)